TTree Multithreaded
It compiles, but we don’t know if it runs or not. Main changes were
making them TTree* inside the forest class rather than shared pointers.
I’m curious to see if this makes a difference as I feel only the TFile
at the end probably needs to be a shared pointer