Environmental Controls
Added the ability to set an environment which will determine which cuts
are performed for particle ID as well as which plots are to be made.
The whole thing then outputs a text file which records what the
settings were. Also Missing mass was updated to have separate plots for
fitting purposes and showing purposes due to the multi-particle
measuring